Querying refactoring start
