+++ ui/src/api/quizApi.ts
@@ -1,0 +1,42 @@
+import {
+  handleError,
+  handleResponseWithJson,
+  handleResponseWithoutJson,
+} from "./apiUtils";
+
+export function saveQuiz(quiz, userId: number) {
+  return fetch(
+    process.env.REACT_APP_FETCH_HOST + "/betterKahoot/quiz/" + userId,
+    {
+      method: "POST",
+      headers: {
+        "Content-Type": "application/json",
+      },
+      body: JSON.stringify(quiz), // body data type must match "Content-Type" header
+    }
+  )
+    .then((response) => handleResponseWithoutJson(response, 201))
+    .catch(handleError);
+}
+
+export function loadAllQuizes(id: number) {
+  return fetch(process.env.REACT_APP_FETCH_HOST + "/betterKahoot/quiz/" + id, {
+    method: "GET",
+    headers: {
+      "Content-Type": "application/json",
+    },
+  })
+    .then((response) => handleResponseWithJson(response, 200))
+    .catch(handleError);
+}
+
+export function deleteQuiz(id: number) {
+  return fetch(process.env.REACT_APP_FETCH_HOST + "/betterKahoot/quiz/" + id, {
+    method: "DELETE",
+    headers: {
+      "Content-Type": "application/json",
+    },
+  })
+    .then((response) => handleResponseWithoutJson(response, 204))
+    .catch(handleError);
+}
