+++ ui/src/common/types.ts
@@ -69,9 +69,3 @@ export interface UserInterface {
   email: string;
   status: string;
 }
-
-export enum ValidationStatus {
-  OK = "OK",
-  TWOANSWERS = "At least 2 answers should be filled",
-  NAMEOFQUESTION = "Name of the question is required",
-}
